requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1
requests==2.32.4
starlette==0.47.3
orjson==3.11.1
typing-extensions==4.14.1

# 数据源
weaviate-client==4.16.9
minio==7.2.16
peewee==3.18.2
sqlalchemy==2.0.42
python-dateutil==2.9.0.post0
validators==0.35.0
requests-toolbelt==1.0.0
tqdm==4.67.1
regex==2025.7.34

# LLM
openai==2.3.0

PyYAML>=6.0
